Rename HAIR to THIR (Typed HIR).
