+++ src/test/ui/pattern/bindings-after-at/borrowck-pat-by-move-and-ref-inverse-promotion.rs
@@ -1,5 +1,5 @@
 // Test that `by_move_binding @ pat_with_by_ref_bindings` is prevented even with promotion.
-// Currently this logic exists in HAIR match checking as opposed to borrowck.
+// Currently this logic exists in THIR match checking as opposed to borrowck.
 
 #![feature(bindings_after_at)]
 #![feature(move_ref_pattern)]
